Moved "movie" guessing into "movie/quality", added "type" filter to "quality" plugin and base searcher methods
